Fix AWS tagging
Tags should be ascii and we shouldn't tag everything as "Environment: Sandbox"
